Fix #8: Add syntax highlighting to input textarea
I couldn't figure out how to load CodeMirror as a module using the
pyscript.json config (everything I tried produced AttributeErrors), so
I added it as a script tag instead.

I'll create an issue for this, but if no one can figure out how to
make it work as a module, the current way should be fine as well.